gan_mnist_project/requirements.txt
tensorflow==2.4.1
numpy==1.19.5
Pillow==8.1.0
